# perf(lsp): use faster version of str_byteindex