perf(lsp): optimize input validation in str_byteindex

The semantic tokens response handler contains a hot loop which makes
lots of calls to vim.str_byteindex which takes most of the time of the
entire handler.

Optimized input validation in vim.str_byteindex. Should be roughly
4x faster.

    name                              total    count      avg
    ────────────────────────────────────────────────
    vim.str_byteindex (original)   635.67ms   263048   2.42us
    vim.str_byteindex (new)        159.20ms   263016   0.61us
    vim._str_byteindex             190.01ms   786164   0.24us