@tris203 2024-11-09
Seems sensible 

I will rebench the lua implementation of this function with the added guard