@lewis6991 2024-11-10
> I will rebench the lua implementation of this function with the added guard

I just tried the Lua version and it was around 10-15% faster. Not sure if it's worth going with.

I also tried using the FFI but that was around the same.

I've guarded some more validation and removed the specialised version. This should be good enough for now.